requirements.txt
protobuf==3.20.0
tensorflow-gpu==2.4.1
keras==2.4.3
numpy==1.19.2
pandas==1.1.5
scipy==1.4.1
Pillow==8.4.0
Cython==0.29.36
matplotlib==3.3.4
scikit-image==0.17.2
imgaug==0.4.0
opencv-python==4.5.5.64
h5py==2.10.0
PyYAML==5.4.1
IPython==7.16.3
git+https://github.com/akTwelve/Mask_RCNN.git#egg=mask-rcnn
